moved reusable scroll view components to components folder
