+++ include/raylibhelper.h
@@ -38,6 +38,8 @@ void draw_trade(PokemonSave *save_player1, PokemonSave *save_player2, char *play
 void draw_file_select_single(struct SaveFileData *save_file_data, PokemonSave *save_player1, char *player1_save_path, struct TrainerInfo *trainer1, struct TrainerSelection *trainer_selection, enum single_player_menu_types menu_type, GameScreen *current_screen);
 void draw_bills_pc(PokemonSave *pkmn_save, char *save_path, struct TrainerInfo *trainer, struct TrainerSelection *trainerSelection, GameScreen *current_screen);
 void draw_evolve(PokemonSave *pkmn_save, char *save_path, struct TrainerInfo *trainer, GameScreen *current_screen);
+void handle_list_scroll(int *y_offset, const int num_saves, const int corrupted_count, int *mouses_down_index, bool *is_moving_scroll, int *banner_position_offset);
+void update_selected_indexes_with_selection(int *selected_saves_index, int *mouses_down_index, bool *is_moving_scroll);
 void draw_raylib_screen_loop(
     struct SaveFileData *save_file_data,
     struct TrainerInfo *trainer1,
