+++ src/raylibhelper.c
@@ -55,6 +55,112 @@ void get_mac_resource_images_path(void)
     }
 }
 #endif
+
+void handle_list_scroll(int *y_offset, const int num_saves, const int corrupted_count, int *mouses_down_index, bool *is_moving_scroll, int *banner_position_offset)
+{
+    const uint8_t box_height = 93;
+    const uint8_t num_visible = 4;
+    const int height = num_saves * box_height - 60 * corrupted_count;
+
+    if (IsMouseButtonDown(MOUSE_LEFT_BUTTON) || IsKeyDown(KEY_UP) || IsKeyDown(KEY_DOWN))
+    {
+        float mouse_delta = GetMouseDelta().y;
+
+        if (mouse_delta >= 0.5 || mouse_delta <= -0.5)
+        {
+            *y_offset += GetMouseDelta().y;
+            *y_offset = *y_offset < -height + (num_visible * box_height) + (corrupted_count * 25) ? -height + (num_visible * box_height) + (corrupted_count * 25) : *y_offset;
+            *y_offset = *y_offset > 75 ? 75 : *y_offset;
+            *mouses_down_index = -1;
+            *is_moving_scroll = true;
+            if (*y_offset < 50)
+            {
+                *banner_position_offset = *y_offset - 50;
+            }
+            else
+            {
+                *banner_position_offset = 0;
+            }
+        }
+        else if (IsKeyDown(KEY_UP))
+        {
+            *y_offset += 1;
+            *mouses_down_index = -1;
+            *is_moving_scroll = true;
+        }
+        else if (IsKeyDown(KEY_DOWN))
+        {
+            *y_offset -= 1;
+            *mouses_down_index = -1;
+            *is_moving_scroll = true;
+        }
+    }
+
+    const int min_y = (height + (num_visible * box_height) + (corrupted_count * 25)) - 75;
+    const int max_y = min_y + 75 - (-height + (num_visible * box_height) + (corrupted_count * 25));
+    const uint8_t t_max = 50;
+    const uint8_t t_min = 0;
+    const uint8_t t_rate = 5;
+    static uint8_t transparency = 0;
+
+    // scroll position indicator
+    float scroll_position = min_y + 75 - *y_offset;
+    scroll_position = (scroll_position - min_y) / (max_y - min_y);
+    scroll_position = scroll_position < 0 ? 0 : scroll_position;
+    scroll_position = scroll_position > 1 ? 1 : scroll_position;
+    if (*is_moving_scroll)
+    {
+        transparency += t_rate;
+        if (transparency > t_max)
+        {
+            transparency = t_max;
+        }
+    }
+    else if (transparency > t_min)
+    {
+        transparency -= t_rate;
+        if (transparency < t_min)
+        {
+            transparency = t_min;
+        }
+    }
+    if (height > SCREEN_HEIGHT - 100 - 2 * (SCREEN_HEIGHT / 16))
+    {
+        DrawRectangleGradientV(SCREEN_WIDTH - 20, 50, 20, SCREEN_HEIGHT / 16, (Color){255, 255, 255, 0}, (Color){255, 255, 255, transparency});
+        DrawRectangle(SCREEN_WIDTH - 20, 50 + SCREEN_HEIGHT / 16, 20, SCREEN_HEIGHT - 100 - 2 * (SCREEN_HEIGHT / 16), (Color){255, 255, 255, transparency});
+        DrawRectangleGradientV(SCREEN_WIDTH - 20, SCREEN_HEIGHT - 50 - SCREEN_HEIGHT / 16, 20, SCREEN_HEIGHT / 16, (Color){255, 255, 255, transparency}, (Color){255, 255, 255, 0});
+        draw_pokeball_scroll(scroll_position, (float)transparency * 5.1f);
+    }
+}
+
+void update_selected_indexes_with_selection(int *selected_saves_index, int *mouses_down_index, bool *is_moving_scroll)
+{
+    if (IsMouseButtonReleased(MOUSE_BUTTON_LEFT))
+    {
+        if (*mouses_down_index != -1 && !*is_moving_scroll)
+        {
+            if (selected_saves_index[0] == *mouses_down_index)
+            {
+                selected_saves_index[0] = -1;
+            }
+            else if (selected_saves_index[1] == *mouses_down_index)
+            {
+                selected_saves_index[1] = -1;
+            }
+            else if (selected_saves_index[0] == -1)
+            {
+                selected_saves_index[0] = *mouses_down_index;
+            }
+            else if (selected_saves_index[1] == -1)
+            {
+                selected_saves_index[1] = *mouses_down_index;
+            }
+        }
+        *mouses_down_index = -1;
+        *is_moving_scroll = false;
+    }
+}
+
 void draw_raylib_screen_loop(
     struct SaveFileData *save_file_data,
     struct TrainerInfo *trainer1,
